in the process of adding longs (push to merge partial seperation)

+++ HPCGLib/include/UtilLib/hpcg_multi_GPU_utils.cuh
@@ -3,7 +3,7 @@
 
 #include "cuda_runtime.h"
 
-typedef int local_int_t;
+typedef long local_int_t;
 typedef int global_int_t;
 
 using DataType = double;
